Refactor convert file into the files related to the conversions
Moves all of the convert.rs contents to specific files versus bundling conversion logic into a singular file that becomes a spaghetti of imports and logic.

1. Conversions to/from `WindowsComponent` have been moved to the associated file.
2. Conversions to/from `UnixComponent` have been moved to the associated file.
